+++ src/app/stores/useProyectsStore/utilities/index.ts
@@ -1,0 +1,2 @@
+// src/app/stores/proyects/utilities/index.ts
+export * from './fetchProyects'